refactor: extract desktop planner and browser session manager
